+++ docs/ReleaseNotes.md
@@ -1,13 +1,25 @@
 # dotnet-httpie release notes
 
-## [0.4.0](https://nuget.org/packages/dotnet-httpie/0.4.0)
+## [0.6.1](https://nuget.org/packages/dotnet-httpie/0.6.1)
+
+- Add `net7.0` support
+- Remove timestamp by default
+- Allow get request with body
+
+## [0.5.3](https://nuget.org/packages/dotnet-httpie/0.5.3)
+
+- Fix `-h` bug
+- Fix download bug
+- Refactor middleware
+
+## [0.4.3](https://nuget.org/packages/dotnet-httpie/0.4.3)
 
 - Add support for download file
 - Add support for response json schema validation
 - Add `RequestCacheMiddleware`
 - Add load test exporter
 
-## [0.3.0](https://nuget.org/packages/dotnet-httpie/0.3.0)
+## [0.3.4](https://nuget.org/packages/dotnet-httpie/0.3.4)
 
 - Add support for load test